requirements.txt
numpy==1.24.2
Pillow==9.5.0
pygame==2.3.0